+++ casper_trade_contracts/bin/build_contract.rs
@@ -2,4 +2,4 @@
 #![no_std]
 #![no_main]
 #![allow(unused_imports, clippy::single_component_path_imports)]
-use casperswap_contracts;
+use casper_trade_contracts;