+++ casper_trade_contracts/src/casper_trade_v2_pair/errors.rs
@@ -1,7 +1,7 @@
 use odra::prelude::OdraError;
 
 #[odra::odra_error]
-pub enum CasperswapV2PairError {
+pub enum CasperTradeV2PairError {
     /// The pair is not properly configured.
     Misconfigured = 20001,
     /// The caller is not allowed to perform this action.