+++ cli/scenarios/add_liquidity.rs
@@ -1,5 +1,5 @@
 use super::utils::{create_token_ref, parse_token_input};
-use casperswap_contracts::router::CasperswapV2Router;
+use casper_trade_contracts::router::CasperTradeV2Router;
 use odra::casper_types::U256;
 use odra::host::HostEnv;
 use odra::prelude::Addressable;
@@ -52,7 +52,7 @@ impl Scenario for AddLiquidity {
         env.set_gas(50_000_000_000);
 
         // Get the router contract
-        let mut router = container.contract_ref::<CasperswapV2Router>(env, None)?;
+        let mut router = container.contract_ref::<CasperTradeV2Router>(env, None)?;
 
         // Get args
         let token_a_input = args.get_single::<String>("token_a")?;