+++ cli/scenarios/add_liquidity_cspr.rs
@@ -1,5 +1,5 @@
 use super::utils::{create_token_ref, parse_token_input};
-use casperswap_contracts::router::CasperswapV2Router;
+use casper_trade_contracts::router::CasperTradeV2Router;
 use odra::casper_types::{U256, U512};
 use odra::host::{HostEnv, HostRef};
 use odra::prelude::Addressable;
@@ -51,7 +51,7 @@ impl Scenario for AddLiquidityCSPR {
         env.set_gas(50_000_000_000);
 
         // Get the router contract
-        let router = container.contract_ref::<CasperswapV2Router>(env, None)?;
+        let router = container.contract_ref::<CasperTradeV2Router>(env, None)?;
 
         // Get args
         let token_a_input = args.get_single::<String>("token_a")?;