+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
-Changelog for `casperswap`.
+Changelog for `Casper Trade`.
 
 ## [0.1.0] - 2025-10-10
 ### Added
-- `casperswap_v2_pair` module.
+- `casper_trade_v2_pair` module.
 - `flapper` module.